services/ml_engine/requirements.txt
scikit-learn==1.3.2
scipy==1.11.4
pandas==2.1.4
numpy==1.26.2
prometheus-api-client==0.5.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1
joblib==1.3.2
schedule==1.2.0

# Optional for advanced models
# torch==2.1.1
# prophet==1.1.5
# river==0.21.0
